Fix bug

+++ src/DataTypes.java
@@ -9,6 +9,6 @@ public class DataTypes {
         for (int x : numbers) {
             s += x;
         }
-        return s;
+        return (long) s;
     }
 }
